Move policy config to PP and out of Mock BPA
