+++ mock-bpa-test/test_bpa.py
@@ -78,13 +78,16 @@ class TestAgent(unittest.TestCase):
         self.assertIsNone(self._ol_sock)
 
         is_json = False
+        use_bcb_rng = False
 
         if testcase is not None:
             policy_config = testcase.policy_config
             LOGGER.info('Using policy config %s', policy_config)
             is_json = policy_config.endswith(".json")
 
+            use_bcb_rng = testcase.use_bcb_rng
             key_set = testcase.key_set
+
         else:
             policy_config = "0x00"
             key_set = "mock-bpa-test/key_set_1.json"
@@ -95,7 +98,8 @@ class TestAgent(unittest.TestCase):
             '-u', 'localhost:4556', '-r', 'localhost:14556',
             '-o', 'localhost:24556', '-a', 'localhost:34556',
             '-j' if is_json else "-p", policy_config,
-            '-k', key_set
+            '-k', key_set,
+            '-c' if use_bcb_rng else ""
         ])
         self._agent = CmdRunner(args, stderr=subprocess.STDOUT)
