+++ src/mock_bpa/key_registry.c
@@ -1,0 +1,133 @@
+/*
+ * Copyright (c) 2025-2026 The Johns Hopkins University Applied Physics
+ * Laboratory LLC.
+ *
+ * This file is part of the Bundle Protocol Security Library (BSL).
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * This work was performed for the Jet Propulsion Laboratory, California
+ * Institute of Technology, sponsored by the United States Government under
+ * the prime contract 80NM0018D0004 between the Caltech and NASA under
+ * subcontract 1700763.
+ */
+
+/** @file
+ * @ingroup mock_bpa
+ */
+
+#include "key_registry.h"
+#include "text_util.h"
+
+int mock_bpa_key_registry_init(const char *pp_cfg_file_path)
+{
+
+    int          retval = 0;
+    json_t      *root;
+    json_error_t err;
+
+    BSL_LOG_INFO("Reading keys from %s", pp_cfg_file_path);
+    root = json_load_file(pp_cfg_file_path, 0, &err);
+    if (!root)
+    {
+        BSL_LOG_ERR("JSON error: line %d: %s", err.line, err.text);
+        json_decref(root);
+        return 1;
+    }
+
+    json_t *keys = json_object_get(root, "keys");
+    if (!keys || !json_is_array(keys))
+    {
+        BSL_LOG_ERR("Missing \"keys\" ");
+        json_decref(root);
+        return 1;
+    }
+
+    size_t n = json_array_size(keys);
+    BSL_LOG_INFO("Found %zu key objects", n);
+
+    for (size_t i = 0; !retval && (i < n); ++i)
+    {
+        json_t *key_obj = json_array_get(keys, i);
+        if (!json_is_object(key_obj))
+        {
+            continue;
+        }
+
+        json_t *kty = json_object_get(key_obj, "kty");
+        if (!kty)
+        {
+            BSL_LOG_ERR("Missing \"kty\" ");
+            continue;
+        }
+
+        if (0 != strcmp("oct", json_string_value(kty)))
+        {
+            BSL_LOG_ERR("Not a symmetric key set");
+            continue;
+        }
+
+        json_t *kid = json_object_get(key_obj, "kid");
+        if (!kid || !json_is_string(kid))
+        {
+            BSL_LOG_ERR("Missing \"kid\" ");
+            continue;
+        }
+        const char *kid_str = json_string_value(kid);
+        BSL_LOG_DEBUG("kid: %s", kid_str);
+
+        json_t *k = json_object_get(key_obj, "k");
+        if (!k || !json_is_string(k))
+        {
+            BSL_LOG_ERR("Missing \"k\" ");
+            continue;
+        }
+        const char *k_str = json_string_value(k);
+        BSL_LOG_DEBUG("k: %s", k_str);
+
+        m_string_t k_text;
+        m_string_init_set_cstr(k_text, k_str);
+        m_bstring_t k_data;
+        m_bstring_init(k_data);
+
+        retval = mock_bpa_base64_decode(k_data, k_text);
+
+        if (!retval)
+        {
+            const size_t   k_len = m_bstring_size(k_data);
+            const uint8_t *k_ptr = m_bstring_view(k_data, 0, k_len);
+
+            retval = BSL_Crypto_AddRegistryKey(kid_str, k_ptr, k_len);
+        }
+        m_bstring_clear(k_data);
+        m_string_clear(k_text);
+    }
+
+    json_decref(root);
+
+    return retval;
+}
+
+int mock_bpa_rfc9173_bcb_cek(unsigned char *buf, int len)
+{
+    if (len == 12) // IV
+    {
+        uint8_t iv[] = { 0x54, 0x77, 0x65, 0x6c, 0x76, 0x65, 0x31, 0x32, 0x31, 0x32, 0x31, 0x32 };
+        memcpy(buf, iv, 12);
+    }
+    else // A3 KEY
+    {
+        uint8_t rfc9173A3_key[] = { 0x71, 0x77, 0x65, 0x72, 0x74, 0x79, 0x75, 0x69,
+                                    0x6f, 0x70, 0x61, 0x73, 0x64, 0x66, 0x67, 0x68 };
+        memcpy(buf, rfc9173A3_key, len);
+    }
+    return 1;
+}
